requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cached-property==1.5.2
Django==4.1.1
django-filter==22.1
django-url-filter==0.3.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
enum-compat==0.0.3
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
